# api-docs.json
{"openapi":"3.0.1","info":{"title":"API de Corte e Religação - OpenMDX","description":"Microsserviço para validação de comandos de corte e religação de energia","contact":{"name":"Equipe OpenMDX","email":"suporte@openmdx.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Comandos","description":"APIs para validação de comandos de corte e religação"},{"name":"Histórico","description":"APIs para consulta de histórico de validações"},{"name":"Feriados","description":"APIs para gerenciamento de feriados"}],"paths":{"/api/holidays":{"get":{"tags":["Feriados"],"summary":"Listar feriados","description":"Retorna todos os feriados cadastrados","operationId":"listarFeriados","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Feriado"}}}}}}},"post":{"tags":["Feriados"],"summary":"Cadastrar feriado","description":"Cadastra um novo feriado nacional ou regional","operationId":"criarFeriado","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeriadoCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeriadoCreateResponse"}}}}}}},"/api/commands/validate":{"post":{"tags":["Comandos"],"summary":"Validar comando","description":"Valida um comando de corte ou religação conforme regras da ANEEL","operationId":"validarComando","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComandoValidateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComandoValidateResponse"}}}}}}},"/api/commands/history":{"get":{"tags":["Histórico"],"summary":"Consultar histórico","description":"Consulta o histórico de validações por UC e/ou período","operationId":"consultarHistorico","parameters":[{"name":"ucId","in":"query","required":false,"schema":{"type":"string"}},{"name":"dataInicio","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"dataFim","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComandoHistoryResponse"}}}}}}}},"components":{"schemas":{"FeriadoCreateRequest":{"type":"object","properties":{"data":{"type":"string","format":"date"},"nome":{"type":"string"},"tipo":{"type":"string"},"regiao":{"type":"string"}}},"FeriadoCreateResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"data":{"type":"string","format":"date"},"nome":{"type":"string"},"tipo":{"type":"string"}}},"ComandoValidateRequest":{"type":"object","properties":{"ucId":{"type":"string"},"tipoUc":{"type":"string"},"tipoComando":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"regiao":{"type":"string"},"solicitante":{"type":"string"}}},"ComandoValidateResponse":{"type":"object","properties":{"aprovado":{"type":"boolean"},"motivo":{"type":"string"},"prazoExecucao":{"type":"string","format":"date-time"}}},"Feriado":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"data":{"type":"string","format":"date"},"nome":{"type":"string"},"tipo":{"type":"string","enum":["nacional","regional"]},"regiao":{"type":"string"}}},"ComandoHistoryItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ucId":{"type":"string"},"tipoComando":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"aprovado":{"type":"boolean"},"motivo":{"type":"string"}}},"ComandoHistoryResponse":{"type":"object","properties":{"total":{"type":"integer","format":"int32"},"comandos":{"type":"array","items":{"$ref":"#/components/schemas/ComandoHistoryItem"}}}}}}}